+++ spec/models/state_file_nc_intake_spec.rb
@@ -184,25 +184,69 @@ RSpec.describe StateFileNcIntake, type: :model do
 
   describe "#calculate_date_electronic_withdrawal" do
     let(:intake) { create(:state_file_nc_intake, :taxes_owed) }
+    let(:state_code) { "nc" }
     let(:timezone) { StateFile::StateInformationService.timezone(state_code) }
-    let(:payment_deadline_date) { StateFile::StateInformationService.payment_deadline_date("nc") }
+    let(:payment_deadline_date) do
+      StateFile::StateInformationService.payment_deadline_date(state_code).in_time_zone(timezone)
+    end
     let(:filing_year) { MultiTenantService.new(:statefile).current_tax_year }
 
     context "when submitted after the payment deadline" do
-      let(:current_time) { payment_deadline_date + 1.day }
-
       it "returns next available date" do
-        expect(intake).to receive(:next_available_date)
-        intake.calculate_date_electronic_withdrawal(current_time: current_time)
+        Time.use_zone(timezone) do
+          Timecop.freeze(payment_deadline_date + 1.hour) do
+            expect(intake).to receive(:next_available_date)
+            intake.calculate_date_electronic_withdrawal(current_time: Time.zone.now)
+          end
+        end
       end
     end
 
     context "when submitted before the payment deadline" do
-      let(:current_time) { payment_deadline_date - 1.day }
-
       it "does not call next_available_date" do
-        expect(intake).not_to receive(:next_available_date)
-        expect(intake.calculate_date_electronic_withdrawal(current_time: current_time)).to eq(DateTime.new(2024, 4, 15))
+        Time.use_zone(timezone) do
+          Timecop.freeze(payment_deadline_date - 1.hour) do
+            expect(intake).not_to receive(:next_available_date)
+            result = intake.calculate_date_electronic_withdrawal(current_time: Time.zone.now)
+            expect(result).to eq(DateTime.new(2024, 4, 15).in_time_zone(timezone))
+          end
+        end
+      end
+    end
+  end
+
+  describe '#next_avaliable_date' do
+    let(:intake) { create :state_file_nc_intake }
+    let(:date) { intake.next_available_date(time) }
+    context "when it is before 5pm and the next day is a valid day" do
+      # 4pm Tuesday
+      let(:time) { DateTime.new(2024, 4, 16, 15, 0, 0)}
+      it "is valid and saves the intake with the next day" do
+        expect(date).to eq(DateTime.new(2024, 4, 17))
+      end
+    end
+
+    context "when it is before 5pm and the next day is a holiday" do
+      # 4pm christmas eve
+      let(:time) { DateTime.new(2024, 12, 24, 15, 0, 0)}
+      it "is valid and saves the intake with a date after the holiday" do
+        expect(date).to eq(DateTime.new(2024, 12, 26))
+      end
+    end
+
+    context "when it is before 5pm and the next day is saturday" do
+      # 4pm friday
+      let(:time) { DateTime.new(2024, 4, 19, 15, 0, 0)}
+      it "is valid and saves the intake with a date after the holiday" do
+        expect(date).to eq(DateTime.new(2024, 4, 22))
+      end
+    end
+
+    context "when it is after 5pm and the next two days are valid" do
+      # 5:30pm Tuesday
+      let(:time) { DateTime.new(2024, 4, 16, 17, 30, 0)}
+      it "is valid and saves the intake with a date 2 business days later" do
+        expect(date).to eq(DateTime.new(2024, 4, 18))
       end
     end
   end
